BIC-192 # move BMP.Forms.getDefinition() declaration to better place - new AMD module provides a function that returns a Promise - Promise is resolved when Forms library is ready to be used
